feat(avo-1932): move render to comp. PupilCollectionForTeacherPreview
