+++ src/assignment/components/AssignmentHeading.scss
@@ -44,7 +44,7 @@
 			align-items: flex-start;
 		}
 
-		&-slim {
+		&__slim {
 			min-height: 0;
 		}
